requirements.txt
Django==5.0.2
mysqlclient==2.2.4
djangorestframework==3.14.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7 